requirements.txt
# AI Shift Agent Requirements
fastapi==0.109.0
jinja2
python-multipart
uvicorn==0.27.0
python-telegram-bot==20.7
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
httpx==0.25.2
pillow==10.2.0
pydantic==2.6.0
pydantic-settings==2.1.0
google-genai

# Database dependencies
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0